Update the documentation
I added return values to the documentation where it was missing - per the CRAN comments.

I also converted the documentation to Markdown (using [roxygen2md](https://roxygen2md.r-lib.org/) as recommended), as it makes the documentation easier to write and maintain.

I also ran `goodpractice::goodpractice()` to check if anything else was possibly missed and this led to a few more cosmetic changes and the removal of `purrr` from Imports as it's not actually being used in the package.