Update Roxygen version
I also followed the advice to remove the deprecated tag `@docType package` and instead document the special object `"_PAKCAGE"`. I used `usethis::use_package_doc()` for this and then moved the `@importFrom`s.

The change for users here is that the top-level documentation available to users in R should be easier to get to (more aliases) and will have better content (since it will now auto-populate from the DESCRIPTION file.