Add tests for percent
Some of these fail but I think they should pass.

I guess we don't necessarily want to depend on `{janitor}` but is there any reason not to copy the code for use here? Then we'd be confident that the functionality matched exactly.

https://github.com/sfirke/janitor/blob/main/R/round_half_up.R

Either way, I think round and signif half-up should be in their own script (and probably split the tests, too) as they're not necessarily related to percent, and we have an option to export them in the future if we want.